Added ChallengeDocumentToDtoConverterTest and testing
