+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/App.java
@@ -1,10 +1,7 @@
 package com.itachallenge.challenge;
 
 import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.data.mongo.MongoDataAutoConfiguration;
-import org.springframework.boot.autoconfigure.mongo.MongoAutoConfiguration;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 
 @SpringBootApplication
@@ -14,4 +11,5 @@ public class App {
     public static void main(String[] args) {
         SpringApplication.run(App.class, args);
     }
+
 }
