--- itachallenge-challenge/src/main/java/com/itachallenge/challenge/config/ModelMapperConfig.java
@@ -1,47 +1,0 @@
-package com.itachallenge.challenge.config;
-
-import org.modelmapper.*;
-import org.modelmapper.convention.MatchingStrategies;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-
-@Configuration
-public class ModelMapperConfig {
-    @Bean
-    public ModelMapper modelMapper() {
-        ModelMapper modelMapper = new ModelMapper();
-
-        Provider<LocalDate> localDateProvider = new AbstractProvider<LocalDate>() {
-            @Override
-            public LocalDate get() {
-                return LocalDate.now();
-            }
-        };
-
-        Converter<String, LocalDate> toStringDate = new AbstractConverter<String, LocalDate>() {
-            @Override
-            protected LocalDate convert(String source) {
-                DateTimeFormatter format = DateTimeFormatter.ofPattern("yyyy-MM-dd");
-                LocalDate localDate = LocalDate.parse(source, format);
-                return localDate;
-            }
-        };
-
-
-
-
-        modelMapper.createTypeMap(String.class, LocalDate.class);
-        modelMapper.addConverter(toStringDate);
-       // modelMapper.getTypeMap(String.class, LocalDate.class).setProvider(localDateProvider);
-
-
-
-
-/*        modelMapper.getConfiguration();
-           // .setMatchingStrategy(MatchingStrategies.LOOSE);*/
-        return new ModelMapper();
-    }
-}
